# docs/api.yaml
swagger: '2.0'
info:
  title: Grakn
  description: 'API for clients to interact with Grakn'
  version: "0.1"
schemes: [http]
host: localhost:4567
paths:
...
  /kb/{keyspace}:
    parameters:
      - $ref: '#/parameters/keyspace'
    get:
      summary: Get Knowledge Base
      tags: [Keyspaces]
      produces: [application/json]
      responses:
        200:
          description: ''
          schema:
            $ref: '#/definitions/KnowledgeBase'
        404:
          description: ''
    put:
      summary: Put Knowledge Base
      description: >-
        If the keyspace exists, nothing happens. The configuration is returned. This response is the same as the
        response from `/kb/configuration`.
      tags: [Keyspaces]
      produces: [application/json]
      responses:
        200:
          description: ''
          schema:
            $ref: '#/definitions/Configuration'
    delete:
      summary: Delete Knowledge Base
      description: >-
        If the keyspace doesn't exist, nothing happens.
      tags: [Keyspaces]
      responses:
        204:
          description: ''
...
parameters:
  keyspace:
    name: keyspace
    in: path
    required: true
    type: string
    x-example: genealogy
#   For OpenAPI 3.0
#   schema:
#     $ref: '#/definitions/Keyspace'
...
definitions:
...
  Keyspace:
    type: string
    example: genealogy
  KnowledgeBase:
    required: ['@id', name, types, roles, rules, graql]
    type: object
    properties:
      '@id':
        type: string
        example: /kb/genealogy
      name:
        $ref: '#/definitions/Keyspace'
      types:
        $ref: '#/definitions/TypesLink'
      roles:
        $ref: '#/definitions/RolesLink'
      rules:
        $ref: '#/definitions/RulesLink'
      graql:
        type: string
        format: uri
        example: /kb/genealogy/graql
...
  TypesLink:
    type: string
    format: uri
    example: /kb/genealogy/type
...
  RolesLink:
    type: string
    format: uri
    example: /kb/genealogy/role
...
  RulesLink:
    type: string
    format: uri
    example: /kb/genealogy/rule
...
  Configuration:
    type: object
#   For OpenAPI 3.0
#   additionalProperties:
#     oneOf:
#       - type: string
#       - type: number
#       - type: boolean
    example:
      knowledge-base.mode: production
      knowledge-base.analytics: distributed
      knowledge-base.default-keyspace: grakn
      knowledge-base.sharding-threshold: 10000
      knowledge-base.schema-cache-timeout-ms: 600000
      server.host: 0.0.0.0
      server.port: 4567
      server.static-file-dir: ../services/assets/
      loader.repeat-commits: 5
      tasks.postprocessing.delay: 60000
      loader.threads: 4
      webserver.threads: 64
      log.dirs: ../logs/
      log.level: INFO
      storage.backend: grakn-production
      storage.hostname: 127.0.0.1
      cache.db-cache: true
      cache.db-cache-clean-wait: 20
      cache.db-cache-time: 180000
      cache.db-cache-size: 0.25
      queue.host: localhost:6379
      queue.consumers: 32
      redis.pool-size: 32
...